pyproject.toml
[project]
name = "pdf-document-extractor"
version = "1.0.0"
description = "Standalone PDF document extraction service using Docling"
authors = [
    { name = "Mark Johnson", email = "mark.johnson@enverus.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI dependencies
    "fastapi~=0.115.14",
    "uvicorn[standard]~=0.35.0",
    "pydantic~=2.10.0",
    "pydantic-settings~=2.10.1",

    # AWS/Cloud dependencies
    "boto3>=1.35.0",
    "boto3-stubs>=1.35.0",

    # Document processing
    "docling>=2.52.0",
    "pypdf>=5.1.0",

    # Utilities
    "httpx>=0.27.2",
    "tenacity>=9.0.0,<10.0.0",
    "lz4>=4.0.0,<5.0.0",
    "toml~=0.10.2",
    "pyyaml>=6.0,<7.0.0",
    "psutil==7.0.0",

    # Observability (optional)
    "opentelemetry-api>=1.36.0,<2.0.0",
    "opentelemetry-sdk>=1.36.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.36.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0,<1.0.0",
    "opentelemetry-instrumentation-requests>=0.57b0,<1.0.0",
    "prometheus-client>=0.21.1,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1,<9.0.0",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "coverage>=7.10.3,<8.0.0",
    "ruff>=0.6.0,<1.0.0",
    "mypy>=1.11.0,<2.0.0",
    "moto[all,s3]~=5.1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "lib"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = false  # Relaxed for faster development
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false  # Relaxed for faster development
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
]
ignore = [
    "E501",   # line too long
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
norecursedirs = ["src", ".git", "dist", "build", "*.egg"]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/models/*",
    "src/configs/settings.py",
    "src/main.py",
    "*/tests/*",
    "*/__init__.py",
]
